# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart University Communication and Venue Notification System API",
    "version": "1.0.0",
    "description": "A comprehensive backend system for university communication, announcements, venue management, and real-time notifications",
    "contact": {
      "name": "University of Zimbabwe",
      "email": "support@uz.ac.zw"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Get course by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Course retrieved"},
          "404": {"description": "Course not found"}
        }
      },
      "put": {
        "tags": ["Courses"],
        "summary": "Update course (Lecturer/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {"type": "string"},
                  "name": {"type": "string"},
                  "description": {"type": "string"},
                  "department": {"type": "string"},
                  "credits": {"type": "integer"},
                  "lecturerId": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Course updated"},
          "403": {"description": "Forbidden"}
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "Delete course (Lecturer/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Course deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
...
    "/api/schedules": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Create schedule (Lecturers/Admins)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["venueId", "courseId", "dayOfWeek", "startTime", "endTime", "semester"],
                "properties": {
                  "venueId": {"type": "string"},
                  "courseId": {"type": "string"},
                  "dayOfWeek": {"type": "string", "enum": ["MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY"]},
                  "startTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "endTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "semester": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Schedule created"},
          "400": {"description": "Time slot already booked"}
        }
      },
      "get": {
        "tags": ["Schedules"],
        "summary": "Get all schedules",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "venueId", "in": "query", "schema": {"type": "string"}},
          {"name": "lecturerId", "in": "query", "schema": {"type": "string"}},
          {"name": "dayOfWeek", "in": "query", "schema": {"type": "string"}},
          {"name": "semester", "in": "query", "schema": {"type": "string"}},
          {"name": "courseId", "in": "query", "schema": {"type": "string"}},
          {"name": "courseCode", "in": "query", "schema": {"type": "string"}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1, "minimum": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20, "minimum": 1}}
        ],
        "responses": {
          "200": {"description": "Schedules retrieved"}
        }
      }
    },
    "/api/schedules/my-schedule": {
      "get": {
        "tags": ["Schedules"],
        "summary": "Get current user's schedule",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "User schedule retrieved"}
        }
      }
    },
    "/api/schedules/{id}": {
      "put": {
        "tags": ["Schedules"],
        "summary": "Update schedule (Author/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "venueId": {"type": "string"},
                  "courseId": {"type": "string"},
                  "dayOfWeek": {"type": "string", "enum": ["MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY"]},
                  "startTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "endTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "semester": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Schedule updated"},
          "403": {"description": "Forbidden"}
        }
      },
      "delete": {
        "tags": ["Schedules"],
        "summary": "Delete schedule (Author/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Schedule deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}